Fix missing token after reload
Fixes the issue that API calls are unauthenticated in case the app is
opened or reloaded although the user is logged in.

The root loader is supposed to provide an instance of the Spotify API
library with an access token set. This change implements the setting
of the token from local storage in case we're not being redirected from
the login.